pyproject.toml
[project]
name = "cyares"
dynamic = ["version"]

description = "A Cython Version of c-ares"
readme = "README.md"
authors = [
    { name = "Vizonex", email = "VizonexBusiness@gmail.com" }
]
requires-python = ">=3.10"

dependencies = [
    # Self argument
    'typing-extensions; python_version<"3.11"',
    
    # yarl support is planned but not ready, 
    # I would like for yarl to have a cimport for URL objects first before
    # I start using it with this project so that maximum performance can
    # be achieved and before that can happen 3 things must take place for 
    # that to be completed. 

    # 1. propcache which yarl utilizes needs to be cimportable,
    #   I have already wrote a pull request that would do that it 
    #   just needs the stamp of approval by another maintainer.
    # 2. multidict needs a CPython capsule, this should be avalible in 6.7
    # 3. yarl will need a cython branch not just for Quoting/UnQuoting but also 
    # for URL objects and an __init__.pxd file for importing with cython
  
]

[tool.setuptools.dynamic]
version = {attr = "cyares.__version__"}


[project.optional-dependencies]
aiohttp = ["aiohttp>=3.12.15"]
all = ["idna>=2.1", "trio>=0.30.0", "aiohttp>=3.12.15", "anyio>=4.12.1"]
idna = ["idna>=2.1"]
trio = ["trio>=0.30.0"]
anyio = ["anyio>=4.12.1"]

[tool.cibuildwheel]
build-frontend = "build"
enable = ["cpython-freethreading"]


[build-system]
requires = ["setuptools", "cython", "wheel"]

[dependency-groups]
dev = [
    "pytest-codspeed>=4.0.0",
]

[tool.setuptools]
zip-safe = false
packages = ['cyares']

[project.urls]
    repository = "https://github.com/Vizonex/cyares.git"


[tool.ruff.lint]
select = [
    "E",
    "W",
    "I",
    "F"
]
ignore = [
    "F405" # may be undefined, or defined from star imports (I Need these)
]
